### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-20007

### Change
- Add isMarketPlace to query getNotification
- Update UI for notification from market place
- Update Notification entity
- Add confirmation popup for delete button
- Update test case for changes

### Test account
- environment: mock
- username: rootadmin

### Output screenshot
 ![Screenshot 2024-03-25 at 16.03.31.png](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/99784/attachments/Screenshot%202024-03-25%20at%2016.03.31.png)
![Screenshot 2024-03-25 at 16.03.41.png](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/99784/attachments/Screenshot%202024-03-25%20at%2016.03.41.png)

EZRX-20007 Change query to add isMarketPlace

+++ lib/presentation/notification/widgets/delete_notifications_button.dart
@@ -16,10 +16,13 @@ class _DeleteNotificationButton extends StatelessWidget {
 
         return IconButton(
           key: WidgetKeys.notificationDeleteButton,
-          onPressed: () {
-            context.read<NotificationBloc>().add(
-                  const NotificationEvent.deleteAllNotifications(),
-                );
+          onPressed: () async {
+            final confirmed = await _showConfirmBottomSheet(context);
+            if ((confirmed ?? false) && context.mounted) {
+              context.read<NotificationBloc>().add(
+                    const NotificationEvent.deleteAllNotifications(),
+                  );
+            }
           },
           icon: const Icon(
             Icons.delete_outline,
@@ -27,6 +30,20 @@ class _DeleteNotificationButton extends StatelessWidget {
           ),
         );
       },
+    );
+  }
+
+  Future<bool?> _showConfirmBottomSheet(BuildContext context) {
+    return showModalBottomSheet<bool>(
+      context: context,
+      enableDrag: false,
+      builder: (_) => ConfirmBottomSheet(
+        key: WidgetKeys.confirmBottomSheet,
+        title: 'Clear all notifications?',
+        content: 'This action cannot be undone',
+        confirmButtonText: 'Clear All',
+        displayCancelButton: true,
+      ),
     );
   }
 }
